More useful primitives for datalog db
Sketch an idea of how storage will be organized.
Equality has importance for joining and querying relations,
hence it's beneficial to get rid of the structure of constants and
store them as Id values in the value domain.
Bdddb uses this for storing relations as binary decision diagrams.
TODO decide which data structure to use, temporary using ArrayList to
get a working semi-naive evaluator first.