Create a namespace for tuple utilities
Moved tuple operations to a sub-namespace primary for an implementation
of Hoare's QuickSelect algorithm that has to be adapted to k-dimensions.

Also rename k-d tree to dynamic k-d tree, remove file operations.
Static k-d tree will be reserved for writing to file and hosting
balanced immutable trees with known size (for evaluated strata).