Add in-memory tuple pointer iterator interface
Export of a static k-d tree requires copying into a memory region to
calculate byte offsets of the node hierarchy. If static tree is being
built from an in-memory source, the only option is to copy an entire
tree into a DirectCollection prior to median sort. But that means
construction needs two separate copies of the data that's already
present in memory. Add a pointer iterator to avoid excessive copies.