Encode parser's state in types
Instead of probing advance() function on a parser instance for new
tokens, make it return a null parser continuation, forcing all rules to
check for EOF (which is null). This approach eliminates runtime
assertions on EOF, makes token annotations more safer to infer,
however makes code too convoluted and calling for a refactoring.